Encapsulate tag state in record data structure
Tag (and Visit) have now become their own data structures, complete with
functions that operate on their state, which simplifies the code and
should make it easier to test. And I really should get around to writing
tests for this code since the codebase, while more organized, is
becoming a bit more complex.

However, I have had trouble running this code as a set of Lambda
functions on AWS. The startup time for some of the Lambda functions is
slow, over 10 seconds long, and the compiled uberjar file is over 27 MB
in size. I don't think this application is a good candidate for Lambda,
although it was fun to give it a try.