Test application code
For a long time I've tried to get away with testing this application
manually, but I plan to start refactoring it extensively, so I want to
have some automated tests in place to catch errors.

But web applications are notoriously hard to test. In the case of this
application, I had to stub out a lot of database-related code, to the
point that a lot of the code that touches the database isn't really
tested at all. I should rewrite those stubs to be actual mocks so that
they make some assertions about the parameters passed into them and how
they are called, but we're getting into the realm of integration tests
anyway, so I'll avoid that work for now.

I also can't test PUT request to the /:tag route, because as far as I
can tell, there isn't a way to make a PUT request with a form-encoded
body using ring-mock, so I'll have to test the code that tests the
handler functions for now instead.

Long story short, these tests aren't perfect, but they're better than
nothing, which is what I had before.