Add the Strict-Transport-Security header to responses
Stack.im is always served over SSL production, and clients should always
make requests over SSL. To communicate this fact, we should send back
the Strict-Transport-Security header. However, this header should only
be sent back in a response over SSL, but we cannot set headers for
SSL-only requests in the proxy server used by our hosting provider.
Thus, we must send it back from Stack.im itself, but only when the
request is an SSL request.

The proxy server in front of Stack.im sends the standard
X-Forwarded-Proto header in cases where it (the proxy server) is
handling an SSL request. If we detect that the forwarded proto is HTTPS,
we should set and return the Strict-Transport-Security header in the
response.