Remove middleware
This was interfering with my ability to test this application on my
local network, for reasons I could not figure out, but it had something
to do with issuing redirects for the canonical host. New plan: in
production, I'm going to run this application behind a Caddy proxy that
will handle the canonical host redirect and setting HSTS headers,
instead of trying to do it in Clojure.